moved MCTGame to interfaces
