+++ src/main/java/org/braekpo1nt/mctmanager/games/interfaces/MCTGame.java
@@ -1,4 +1,4 @@
-package org.braekpo1nt.mctmanager.games;
+package org.braekpo1nt.mctmanager.games.interfaces;
 
 import org.bukkit.entity.Player;
